Handle user registration POST request.
This change addresses the need by validating and building a query to send to the user registration data base. Database is accessed in the utilities folder. Query will send back a 201 status code if successfully registered in the database and a 400 response if username or email is invalid/already exists or other error occured.