Implement forecast endpoint.
This change addresses the need by first building the endpoint to hit,
then using the 'request' module to retrieve the forecast data. The data
that is sent back is a list of forecasts. The list has a forecast for
every 3 hours. Since we are returning 5 forecasts total then the program
will iterate every 8 forecasts(every 24 hours) and extract the min
temperature, max temperature, and a weather description for the day.
There is also a partition function used for partitioning the array into
lists of size 8.