Fix 24 hourly forecast endpoint.
This change addresses the need by changing the weather API endpoint to
use the One Time call. Should give a list of hourly forecasts. The
endpoint will then iterate through the first 24 of them and get the
date, temperature, and weather description. Package the results and send
back to client.