Add endpoint for verifying user.
Why: This endpoint will allow the client to call this endpoint when the user has verified their email. It will set the verification field to 1.
On the client side when the user is in the verification page and they enter the code and press submit, if the verification code matches than the client will hit this endpoint so that the user is now verified.